Use middleware to add nonce and CSP headers
