+++ src/components/uswds/Identifier.tsx
@@ -1,4 +1,4 @@
-import Image from 'next/image';
+import Image from '@/components/Image';
 
 export function Identifier() {
   const links = [
