+++ src/middlewares/types.ts
@@ -1,0 +1,4 @@
+import { NextMiddleware } from 'next/server';
+
+// eslint-disable-next-line no-unused-vars
+export type MiddlewareFactory = (middleware: NextMiddleware) => NextMiddleware;
